requirements.txt
spacy==3.0.6
pt-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_sm-3.0.0/pt_core_news_sm-3.0.0-py3-none-any.whl
APScheduler==3.6.3
certifi==2020.12.5
python-dotenv==0.17.1
python-telegram-bot==13.5
pytz==2021.1
six==1.16.0
tornado==6.1
tzlocal==2.1
nltk==3.6.2
scikit-learn==0.24.2
